Audit data-display agnostic tests
Entire-Checkpoint: c2b6798ef929
