+++ crates/ars-components/tests/proptest_state_machines.rs
@@ -14,7 +14,7 @@ mod input;
 #[path = "proptest_state_machines/date_time.rs"]
 mod date_time;
 
-#[path = "proptest_state_machines/data_display.rs"]
+#[path = "proptest_state_machines/data_display/mod.rs"]
 mod data_display;
 
 #[path = "proptest_state_machines/layout.rs"]
