+++ crates/ars-components/tests/spec_conformance.rs
@@ -17,7 +17,7 @@
 #[path = "spec_conformance/helper.rs"]
 mod helper;
 
-#[path = "spec_conformance/data_display.rs"]
+#[path = "spec_conformance/data_display/mod.rs"]
 mod data_display;
 
 #[path = "spec_conformance/date_time.rs"]
