+++ crates/ars-components/tests/spec_conformance/data_display/stat.rs
@@ -1,0 +1,18 @@
+use ars_components::data_display::stat;
+
+use super::*;
+
+#[test]
+fn stat_anatomy_matches_spec() {
+    assert_anatomy(
+        "stat",
+        &[
+            (stat::Part::Root, "root"),
+            (stat::Part::Label, "label"),
+            (stat::Part::Value, "value"),
+            (stat::Part::Change, "change"),
+            (stat::Part::TrendIndicator, "trend-indicator"),
+            (stat::Part::HelpText, "help-text"),
+        ],
+    );
+}
